latest 16-5

+++ main.py
@@ -1,37 +1,31 @@
 from operator import contains
 from models import *
-from peewee import DoesNotExist
 
 __winc_id__ = "d7b474e9b3a54d23bca54879a4f1855b"
 __human_name__ = "Betsy Webshop"
- 
+
 
 def search(term):
-    while True:
-
-        try:    
-            products = []
-            term = term.lower()
-            query = (Product.select()
-                    .where(fn.LOWER(Product.name).contains(term.lower()) or fn.LOWER(Product.description).contains(term.lower())))
-            products = [prod for prod in query]
-            if not products:
-                return print(f'{term} was not found!') 
-            else:
-                return products
-        except ValueError:
-            print('No term supplied')
-
-    
-    
+    products = []
+    term = term.lower()
+    query = (Product.select()
+            .where(fn.LOWER(Product.name).contains(term.lower()) or fn.LOWER(Product.description).contains(term.lower())))
+    products = [prod for prod in query]
+    if not products:
+        print(f'{term} was not found!')
+    return products
 
 
 def list_user_products(user):
+    if user is None:
+        return print(f'User was not found!')
     user_products = []
     query = (Product.select()
-             .where(Product.owner == user))
+            .where(Product.owner == user))
     user_products = [prod for prod in query]
-    return user_products
+    if user_products:
+        return print(user_products)
+    return print(f'No products found for {user.name}')
 
 
 def list_products_per_tag(tag):
@@ -40,19 +34,22 @@ def list_products_per_tag(tag):
              .join(Tags, on=(Product.id == Tags.id))
              .where(fn.LOWER(Product.name).contains(fn.LOWER(tag))))
     tag_products = [prod for prod in query]
-    return tag_products
+    if tag_products:
+        return print(tag_products)
+    return print(f'No products with the name {tag} found')
 
 
-def add_product_to_catalog(user_id, product):
-    # query = (User.select()
-    #          .where(User.id == user_id))
-    # new_prod = query.insert(product)
-    # new_prod.execute()
-    # new_prod = new_prod.update(new_prod.owner == user_id)
-    query = (Product.create(product))
-    # query.update(query.owner == user_id)
-    # query.execute()
-    return print(f'Product with {product.name} added to catalog')
+def add_product_to_catalog(user_id: int, product: Product):
+    user = (User.select()
+            .where(User.id == user_id).first())
+    if user:
+        product.owner = user
+        product.save()
+        print(f'Product with {product.name} added to catalog of {user.name}')
+        return True
+    else:
+        print(f'{user_id} does not exist')
+        return False
 
 
 def update_stock(product_id, new_quantity):
@@ -67,11 +64,13 @@ def purchase_product(product_id, buyer_id, quantity):
 
 
 def remove_product(product_id):
+    if product_id is None or (Product.select().where(Product.id == product_id).first()) is None:
+        return 'That product ID is not valid'
     query = (Product.delete()
              .where(Product.id == product_id))
     query.execute()
+    return f'Product with ID {product_id} removed'
 
-    return print(f'Product with ID {product_id} removed')
 
 def populate_test_database():
     #Tags
@@ -89,8 +88,10 @@ def populate_test_database():
     # Users
     jasper = User(name='Jasper', address='street 5, 1234ab, amsterdam', billing='street 5, 1234ab, amsterdam')
     piet = User(name='Piet', address='lane 8, 1234cd, kudel', billing='lane 8, 1234cd, kudel')
+    koos = User(name='Koos', address='lane 9, 1234cd, kudel', billing='lane 9, 1234cd, kudel')
     jasper.save()
     piet.save()
+    koos.save()
 
     # Products
     small_white_fridge = Product(name='small_white_fridge', description='small white fridge', price_per_unit=10, qty_in_stock=5, owner=jasper)
@@ -116,18 +117,19 @@ def populate_test_database():
     # large_black_fridge.save()
 
 
-
-
 populate_test_database()
 
 user1 = User.select().where(User.name == 'Jasper').first()
+user3 = User.select().where(User.name == 'Koos').first()
 tag1 = 'fRidge'
-prod1 = Product(name='small_black_fridge', description='small black fridge', price_per_unit=11, qty_in_stock=3)
+prod1 = Product.select().where(Product.name == 'small_white_fridge').first()
 prod2 = Product.select().where(Product.name == 'large_black_fridge').first()
-print(search())
-#print(list_user_products(user1))
-#print(list_products_per_tag(tag1))
-#print(add_product_to_catalog(user1.id, prod1))
+prod3 = Product(name='small_black_fridge', description='small black fridge', price_per_unit=11, qty_in_stock=3)
+#search('fridgsde')
+#list_user_products(user3)
+#list_products_per_tag(tag1)
+#add_product_to_catalog(1, prod3)
 #print(update_stock(prod1.id, 2))
 #print(purchase_product(prod2.id, user1.id, 1))
 #print(remove_product(prod1.id))
+#remove_product(5)
